mentorhub/requirements.txt
# Python packages required for CodeMentorHub
# Installation: pip install -r requirements.txt

Django==5.0.1
django-crispy-forms==2.0
crispy-bootstrap5==0.7
Pillow==10.0.0

# Email and authentication
django-allauth==0.57.0

# Google Meet and Calendar Integration (optional but recommended)
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
google-auth==2.27.0

# Database
psycopg2-binary==2.9.9  # For PostgreSQL (optional)

# Production server
gunicorn==21.2.0
whitenoise==6.6.0

# Environment variables
python-dotenv==1.0.0

# Testing
pytest==7.4.0
pytest-django==4.7.0
